Asynchronous saving of the recording, extended  recording buffer up to 4 GB, and bugs fixed in encoding.
git-svn-id: svn://svn.code.sf.net/p/glmixer/Source/trunk@1066 a4deb95c-d308-4ec4-8dc1-6419bbeeb61a

+++ src/common.cpp
@@ -182,7 +182,7 @@ void addPathToSystemPath(QByteArray path)
 }
 
 
-QString getByteSizeString(float numbytes)
+QString getByteSizeString(double numbytes)
 {
     QStringList list;
     list << "KB" << "MB" << "GB" << "TB";
