+++ src/common.h
@@ -65,6 +65,6 @@ int intFromBlendingPreset(GLenum, GLenum);
 
 void addPathToSystemPath(QByteArray path);
 
-QString getByteSizeString(float numbytes);
+QString getByteSizeString(double numbytes);
 
 #endif /*  COMMON_H_ */
